Replace domain bootstrapper with a function and improve its API
